Create a new gui module and rearrange the event manager
