+++ app/hypnos/include/gfx.h
@@ -22,22 +22,12 @@ enum battery_symbol {
 	BAT_EMPTY
 };
 
-enum screen {
-	WATCH,
-	INFO
-};
-
-struct gui {
-	enum screen sc;
-};
-
 void gfx_init(void);
 void gfx_battery_set_label(enum battery_symbol);
 void gfx_bt_set_label(enum bt_symbol);
 void gfx_time_set_label(char *);
 void gfx_date_set_label(char *);
 void gfx_update(void);
-void gfx_gui_set_screen(struct gui*, enum screen);
 void gfx_show_info(void);
 void gfx_show_date(void);
